// src/domain/enums/status-code.ts
export enum StatusCode {
    ok = 200,
    created = 201,
    noData = 204,
    badRequest = 400,
    unauthorized = 401,
    forbidden = 403,
    notFound = 404,
    methodNotAllowed = 405,
    notAcceptable = 406,
    conflict = 409,
    invalidMediaType = 415,
    internalServerError = 500,
    notImplemented = 501,
    badGateway = 502,
}